requirements.txt
# RAG Application Requirements

# Core libraries
streamlit>=1.22.0
streamlit-chat>=0.0.2.2

# OpenAI
openai>=1.3.5

# LangChain libraries
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.8

# Vector database
faiss-cpu>=1.7.4

# Document processing
pypdf2>=3.0.0
docx2txt>=0.8
tiktoken>=0.5.0

# For deployment
python-dotenv>=1.0.0